requirements.txt
flask
flask-sqlalchemy
flask-login
pdfplumber
PyPDF2
python-docx
scikit-learn
groq
python-dotenv
werkzeug